Simplify Documenter classes for params and example
Removed the ShpaeDocumenter class and made it a method that dispatches a
documenter's method. Then moved the ExampleDocumenters and ParamsDocumenters
to their own modules.
